organize workers/go, add apps/ dir for app entrypoints, workerlib/ for shared worker code (i.e. workflows, activities, etc.)
