+++ workers/go/apps/worker/worker.go
@@ -10,10 +10,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/secretsmanager"
 	"github.com/nexus-rpc/sdk-go/nexus"
 	"github.com/temporalio/omes/cmd/clioptions"
-	"github.com/temporalio/omes/workers/go/ebbandflow"
 	"github.com/temporalio/omes/workers/go/harness"
-	"github.com/temporalio/omes/workers/go/kitchensink"
-	"github.com/temporalio/omes/workers/go/schedulerstress"
+	"github.com/temporalio/omes/workers/go/workerlib/ebbandflow"
+	"github.com/temporalio/omes/workers/go/workerlib/kitchensink"
+	"github.com/temporalio/omes/workers/go/workerlib/schedulerstress"
 	"go.temporal.io/sdk/activity"
 	sdkclient "go.temporal.io/sdk/client"
 	"go.temporal.io/sdk/contrib/aws/lambdaworker"
